backend/api/requirements.txt
python-dotenv
flask
flask-mysqldb
pytest
pytest-mock
flask-jwt-extended
bcrypt
PyJWT